Created a new module that can take populater's place using a different generation method
To test the flexibility of loading in different level generation methods, I created rogueLevel.lua which can swap in for populater.lua as populateMap within level.lua.

Unfortunately, this revealed a flaw in that the level generation methods could all export using the callback system but are missing a uniform syntax to export through a copy of self.

defineLevel removed as it is no longer as useful as it was during the refactoring when it was made.